Enum kate::structs::loaded::constant_pool::Data

pub enum Data {

    Utf8(Arc<Utf8Data>),
    Integer(IntegerData),
    Float(FloatData),
    Long(LongData),
    Double(DoubleData),
    Class(Arc<ClassData>),
    String(StringData),
    FieldRef(Arc<FieldRefData>),
    MethodRef(MethodRefData),
    InterfaceMethodRef(InterfaceMethodRefData),
    NameAndType(Arc<NameAndTypeData>),
    MethodHandle(MethodHandleData),
    MethodType(MethodTypeData),
    Dynamic(DynamicData),
    InvokeDynamic(InvokeDynamicData),
    Module(ModuleData),
    Package(PackageData),

}

Variants

Utf8(Arc<Utf8Data>)

Integer(IntegerData)

Float(FloatData)

Long(LongData)

Double(DoubleData)

Class(Arc<ClassData>)

String(StringData)

FieldRef(Arc<FieldRefData>)

MethodRef(MethodRefData)

InterfaceMethodRef(InterfaceMethodRefData)

NameAndType(Arc<NameAndTypeData>)

MethodHandle(MethodHandleData)

MethodType(MethodTypeData)

Dynamic(DynamicData)

InvokeDynamic(InvokeDynamicData)

Module(ModuleData)

Package(PackageData)

Implementations

impl Data

pub fn as_utf8_mut(&mut self) -> Option<&mut Arc<Utf8Data>>

Optionally returns mutable references to the inner fields if this is a Data::Utf8, otherwise None

pub fn as_utf8(&self) -> Option<&Arc<Utf8Data>>

Optionally returns references to the inner fields if this is a Data::Utf8, otherwise None

pub fn into_utf8(self) -> Result<Arc<Utf8Data>, Self>

Returns the inner fields if this is a Data::Utf8, otherwise returns back the enum in the Err case of the result

pub fn as_integer_mut(&mut self) -> Option<&mut IntegerData>

Optionally returns mutable references to the inner fields if this is a Data::Integer, otherwise None

pub fn as_integer(&self) -> Option<&IntegerData>

Optionally returns references to the inner fields if this is a Data::Integer, otherwise None

pub fn into_integer(self) -> Result<IntegerData, Self>

Returns the inner fields if this is a Data::Integer, otherwise returns back the enum in the Err case of the result

pub fn as_float_mut(&mut self) -> Option<&mut FloatData>

Optionally returns mutable references to the inner fields if this is a Data::Float, otherwise None

pub fn as_float(&self) -> Option<&FloatData>

Optionally returns references to the inner fields if this is a Data::Float, otherwise None

pub fn into_float(self) -> Result<FloatData, Self>

Returns the inner fields if this is a Data::Float, otherwise returns back the enum in the Err case of the result

pub fn as_long_mut(&mut self) -> Option<&mut LongData>

Optionally returns mutable references to the inner fields if this is a Data::Long, otherwise None

pub fn as_long(&self) -> Option<&LongData>

Optionally returns references to the inner fields if this is a Data::Long, otherwise None

pub fn into_long(self) -> Result<LongData, Self>

Returns the inner fields if this is a Data::Long, otherwise returns back the enum in the Err case of the result

pub fn as_double_mut(&mut self) -> Option<&mut DoubleData>

Optionally returns mutable references to the inner fields if this is a Data::Double, otherwise None

pub fn as_double(&self) -> Option<&DoubleData>

Optionally returns references to the inner fields if this is a Data::Double, otherwise None

pub fn into_double(self) -> Result<DoubleData, Self>

Returns the inner fields if this is a Data::Double, otherwise returns back the enum in the Err case of the result

pub fn as_class_mut(&mut self) -> Option<&mut Arc<ClassData>>

Optionally returns mutable references to the inner fields if this is a Data::Class, otherwise None

pub fn as_class(&self) -> Option<&Arc<ClassData>>

Optionally returns references to the inner fields if this is a Data::Class, otherwise None

pub fn into_class(self) -> Result<Arc<ClassData>, Self>

Returns the inner fields if this is a Data::Class, otherwise returns back the enum in the Err case of the result

pub fn as_string_mut(&mut self) -> Option<&mut StringData>

Optionally returns mutable references to the inner fields if this is a Data::String, otherwise None

pub fn as_string(&self) -> Option<&StringData>

Optionally returns references to the inner fields if this is a Data::String, otherwise None

pub fn into_string(self) -> Result<StringData, Self>

Returns the inner fields if this is a Data::String, otherwise returns back the enum in the Err case of the result

pub fn as_field_ref_mut(&mut self) -> Option<&mut Arc<FieldRefData>>

Optionally returns mutable references to the inner fields if this is a Data::FieldRef, otherwise None

pub fn as_field_ref(&self) -> Option<&Arc<FieldRefData>>

Optionally returns references to the inner fields if this is a Data::FieldRef, otherwise None

pub fn into_field_ref(self) -> Result<Arc<FieldRefData>, Self>

Returns the inner fields if this is a Data::FieldRef, otherwise returns back the enum in the Err case of the result

pub fn as_method_ref_mut(&mut self) -> Option<&mut MethodRefData>

Optionally returns mutable references to the inner fields if this is a Data::MethodRef, otherwise None

pub fn as_method_ref(&self) -> Option<&MethodRefData>

Optionally returns references to the inner fields if this is a Data::MethodRef, otherwise None

pub fn into_method_ref(self) -> Result<MethodRefData, Self>

Returns the inner fields if this is a Data::MethodRef, otherwise returns back the enum in the Err case of the result

pub fn as_interface_method_ref_mut(
&mut self
) -> Option<&mut InterfaceMethodRefData>

Optionally returns mutable references to the inner fields if this is a Data::InterfaceMethodRef, otherwise None

pub fn as_interface_method_ref(&self) -> Option<&InterfaceMethodRefData>

Optionally returns references to the inner fields if this is a Data::InterfaceMethodRef, otherwise None

pub fn into_interface_method_ref(self) -> Result<InterfaceMethodRefData, Self>

Returns the inner fields if this is a Data::InterfaceMethodRef, otherwise returns back the enum in the Err case of the result

pub fn as_name_and_type_mut(&mut self) -> Option<&mut Arc<NameAndTypeData>>

Optionally returns mutable references to the inner fields if this is a Data::NameAndType, otherwise None

pub fn as_name_and_type(&self) -> Option<&Arc<NameAndTypeData>>

Optionally returns references to the inner fields if this is a Data::NameAndType, otherwise None

pub fn into_name_and_type(self) -> Result<Arc<NameAndTypeData>, Self>

Returns the inner fields if this is a Data::NameAndType, otherwise returns back the enum in the Err case of the result

pub fn as_method_handle_mut(&mut self) -> Option<&mut MethodHandleData>

Optionally returns mutable references to the inner fields if this is a Data::MethodHandle, otherwise None

pub fn as_method_handle(&self) -> Option<&MethodHandleData>

Optionally returns references to the inner fields if this is a Data::MethodHandle, otherwise None

pub fn into_method_handle(self) -> Result<MethodHandleData, Self>

Returns the inner fields if this is a Data::MethodHandle, otherwise returns back the enum in the Err case of the result

pub fn as_method_type_mut(&mut self) -> Option<&mut MethodTypeData>

Optionally returns mutable references to the inner fields if this is a Data::MethodType, otherwise None

pub fn as_method_type(&self) -> Option<&MethodTypeData>

Optionally returns references to the inner fields if this is a Data::MethodType, otherwise None

pub fn into_method_type(self) -> Result<MethodTypeData, Self>

Returns the inner fields if this is a Data::MethodType, otherwise returns back the enum in the Err case of the result

pub fn as_dynamic_mut(&mut self) -> Option<&mut DynamicData>

Optionally returns mutable references to the inner fields if this is a Data::Dynamic, otherwise None

pub fn as_dynamic(&self) -> Option<&DynamicData>

Optionally returns references to the inner fields if this is a Data::Dynamic, otherwise None

pub fn into_dynamic(self) -> Result<DynamicData, Self>

Returns the inner fields if this is a Data::Dynamic, otherwise returns back the enum in the Err case of the result

pub fn as_invoke_dynamic_mut(&mut self) -> Option<&mut InvokeDynamicData>

Optionally returns mutable references to the inner fields if this is a Data::InvokeDynamic, otherwise None

pub fn as_invoke_dynamic(&self) -> Option<&InvokeDynamicData>

Optionally returns references to the inner fields if this is a Data::InvokeDynamic, otherwise None

pub fn into_invoke_dynamic(self) -> Result<InvokeDynamicData, Self>

Returns the inner fields if this is a Data::InvokeDynamic, otherwise returns back the enum in the Err case of the result

pub fn as_module_mut(&mut self) -> Option<&mut ModuleData>

Optionally returns mutable references to the inner fields if this is a Data::Module, otherwise None

pub fn as_module(&self) -> Option<&ModuleData>

Optionally returns references to the inner fields if this is a Data::Module, otherwise None

pub fn into_module(self) -> Result<ModuleData, Self>

Returns the inner fields if this is a Data::Module, otherwise returns back the enum in the Err case of the result

pub fn as_package_mut(&mut self) -> Option<&mut PackageData>

Optionally returns mutable references to the inner fields if this is a Data::Package, otherwise None

pub fn as_package(&self) -> Option<&PackageData>

Optionally returns references to the inner fields if this is a Data::Package, otherwise None

pub fn into_package(self) -> Result<PackageData, Self>

Returns the inner fields if this is a Data::Package, otherwise returns back the enum in the Err case of the result